chore: data migration support specify max speed

+++ include/common/tmsg.h
@@ -875,7 +875,7 @@ int32_t tDeserializeSTrimDbReq(void* buf, int32_t bufLen, STrimDbReq* pReq);
 
 typedef struct {
   int64_t timestamp;  // unit: millisecond
-  int32_t maxSpeed;   // 0 no limit, unit: bit/s
+  int64_t maxSpeed;   // 0 no limit, unit: Byte/s
 } SVTrimDbReq;
 
 int32_t tSerializeSVTrimDbReq(void* buf, int32_t bufLen, SVTrimDbReq* pReq);
